Graph Visualization final - Cleaned extra files & code
