+++ src/client/src/components/GraphVisualisation.vue
@@ -34,77 +34,170 @@ onMounted(() => {
     .append('g')
     .attr('transform', 'translate(40,0)')
 
+  // Add arrowhead marker
+  svg
+    .append('defs')
+    .append('marker')
+    .attr('id', 'arrow')
+    .attr('viewBox', '0 -5 10 10')
+    .attr('refX', 16)
+    .attr('refY', 0)
+    .attr('markerWidth', 6)
+    .attr('markerHeight', 6)
+    .attr('orient', 'auto-start-reverse')
+    .attr('markerUnits', 'strokeWidth')
+    .append('path')
+    .attr('d', 'M0,-5L10,0L0,5')
+    .attr('fill', 'black')
+
   function zoomed(event: any) {
     svg.attr('transform', event.transform)
   }
 
   const root = d3.hierarchy(data)
 
-  const treeLayout = d3.tree().nodeSize([nodeDistance, 300])
-  treeLayout(root as unknown as d3.HierarchyNode<unknown>)
+  function update(source: any) {
+    // Assigns the x and y position for the nodes
+    const treeLayout = d3.tree().nodeSize([nodeDistance, 300])
+    const treeData = treeLayout(root as unknown as d3.HierarchyNode<unknown>)
 
-  const nodeById: { [key: string]: any } = {}
-  root.each((d) => {
-    nodeById[d.data.name] = d
-  })
+    // Compute the new tree layout.
+    const nodes = treeData.descendants()
+    const links = treeData.links()
 
-  const linkPathGenerator: any = d3
+    // Normalize for fixed-depth.
+    nodes.forEach((d) => (d.y = d.depth * 180))
+
+    // ********** Nodes section **********
+
+    const node = svg.selectAll('g.node').data(nodes, (d: any) => d.id || (d.id = d.data.name))
+
+    // Enter any new nodes at the parent's previous position.
+    const nodeEnter: any = node
+      .enter()
+      .append('g')
+      .attr('class', 'node')
+      .attr('transform', () => `translate(${source.y0},${source.x0})`)
+      .on('click', (event, d) => toggleCollapse(d))
+      .style('cursor', (d: any) => (d.children || d._children ? 'pointer' : 'default'))
+
+    // Add Circle for the nodes
+    nodeEnter
+      .append('circle')
+      .attr('r', 5)
+      .style('fill', (d: any) => (d._children ? 'lightsteelblue' : '#999'))
+
+    // Add labels for the nodes
+    nodeEnter
+      .append('text')
+      .attr('dy', '.35em')
+      .attr('x', (d: any) => (d.children || d._children ? -10 : 10))
+      .style('text-anchor', (d: any) => (d.children || d._children ? 'end' : 'start'))
+      .text((d: any) => d.data.name)
+
+    // Update the node positions
+    const nodeUpdate = nodeEnter.merge(node)
+
+    nodeUpdate.attr('transform', (d: any) => `translate(${d.y},${d.x})`)
+
+    // Update the node attributes and style
+    nodeUpdate
+      .select('circle')
+      .attr('r', 5)
+      .style('fill', (d: any) => (d._children ? 'lightsteelblue' : '#999'))
+
+    // Remove any exiting nodes
+    const nodeExit = node.exit().remove()
+
+    nodeExit.select('circle').attr('r', 0)
+
+    nodeExit.select('text').style('fill-opacity', 0)
+
+    // ********** Links section **********
+
+    const link: any = svg.selectAll('path.link').data(links, (d: any) => d.target.id)
+
+    // Enter new links at the parent's previous position.
+    const linkEnter = link
+      .enter()
+      .insert('path', 'g')
+      .attr('class', 'link')
+      .attr('d', () => {
+        const o = { x: source.x0, y: source.y0 }
+        return diagonal({ source: o, target: o })
+      })
+      .attr('stroke', 'black')
+      .attr('fill', 'none')
+      .attr('marker-end', 'url(#arrow)') // Added attribute for arrow head
+
+    // Update links
+    const linkUpdate = linkEnter.merge(link)
+
+    linkUpdate.attr('d', diagonal)
+
+    // Remove exiting links
+    link
+      .exit()
+      .attr('d', () => {
+        const o = { x: source.x, y: source.y }
+        return diagonal({ source: o, target: o })
+      })
+      .remove()
+
+    const extraLinks: any = [];
+    nodes.forEach((d: any) => {
+      if (d.data.extra_parents) {
+        d.data.extra_parents.forEach((parent: any) => {
+          const parentNode = nodes.find((node: any) => node.data.name === parent.name);
+          if (parentNode) {
+            extraLinks.push({source: parentNode, target: d});
+          }
+        });
+      }
+    });
+
+    const extraLink = svg.selectAll('path.extra-link').data(extraLinks);
+
+    extraLink
+      .enter()
+      .insert('path', 'g')
+      .attr('class', 'extra-link')
+      .attr('stroke', 'red')
+      .attr('fill', 'none')
+      .attr('marker-end', 'url(#arrow)') // Added attribute for arrow head
+      .attr('d', (d: any) => diagonal({source: d.source, target: d.target}));
+
+    extraLink.merge(extraLink).attr('d', (d: any) => diagonal({ source: d.source, target: d.target}));
+
+      //Remove any exiting extra links
+    extraLink.exit().remove();
+
+    // Store the old positions for transition.
+    nodes.forEach((d: any) => {
+      d.x0 = d.x
+      d.y0 = d.y
+    })
+  }
+
+  // Toggle children on click.
+  function toggleCollapse(d: any) {
+    if (d.children) {
+      d._children = d.children
+      d.children = null
+    } else {
+      d.children = d._children
+      d._children = null
+    }
+    update(d)
+  }
+
+  const diagonal: any = d3
     .linkHorizontal()
     .x((d: any) => d.y)
     .y((d: any) => d.x)
 
-  svg
-    .selectAll('path.link')
-    .data(root.links())
-    .enter()
-    .append('path')
-    .attr('class', 'link')
-    .attr('d', linkPathGenerator)
-    .attr('stroke', 'black')
-    .attr('fill', 'none')
-
-  svg
-    .selectAll('g.node')
-    .data(root.descendants())
-    .enter()
-    .append('g')
-    .attr('class', 'node')
-    .attr('transform', (d) => `translate(${d.y},${d.x})`)
-    .append('circle')
-    .attr('r', 5)
-
-  svg
-    .selectAll('g.node')
-    .append('text')
-    .attr('dy', '.35em')
-    .attr('x', (d: any) => (d.children ? -10 : 10))
-    .style('text-anchor', (d: any) => (d.children ? 'end' : 'start'))
-    .text((d: any) => d.data.name)
-    .attr('stroke', 'white')
-    .attr('paint-order', 'stroke')
-
-  const extraLinks: Array<any> = []
-
-  root.descendants().forEach((d) => {
-    if (d.data.extra_parents) {
-      d.data.extra_parents.forEach((parent: { name: string | number }) => {
-        const parentNode = nodeById[parent.name]
-        if (parentNode) {
-          extraLinks.push({ source: parentNode, target: d })
-        }
-      })
-    }
-  })
-
-  svg
-    .selectAll('path.extra-link')
-    .data(extraLinks)
-    .enter()
-    .append('path')
-    .attr('class', 'link') // Same style as the normal links
-    .attr('d', (d) => linkPathGenerator({ source: d.source, target: d.target }))
-    .attr('stroke', 'black')
-    .attr('fill', 'none')
+  // Start the rendering
+  update(root)
 
   const endTime = performance.now()
   console.log(`Rendering took ${endTime - startTime} ms`)
@@ -112,10 +205,6 @@ onMounted(() => {
 </script>
 
 <style scoped>
-body {
-  font-family: Arial, sans-serif;
-}
-
 .node circle {
   fill: #999;
 }
